use-case-demos/finance-trino-demo/requirements.txt
pyarrow==23.0.1
pyiceberg[s3fs]==0.11.1
pyyaml==6.0.3
